COLLECTOR-6131. Improve finally clauses for PostgreSQL CDC tests
Reviewed-on: https://review.streamsets.net/c/datacollector-tests/+/78392
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>

+++ stage/test_postgres_aurora_cdc.py
@@ -234,6 +234,7 @@ def test_stop_start(sdc_builder,
                                                sqlalchemy.String(20)))
     replication_slot = get_random_string(string.ascii_lowercase, 10)
 
+    pipeline = None
     try:
         pipeline_builder = sdc_builder.get_pipeline_builder()
         aurora_postgresql_cdc_client = pipeline_builder.add_stage('Aurora PostgreSQL CDC Client')
@@ -324,7 +325,8 @@ def test_stop_start(sdc_builder,
             records = [{'id': record.field['id'], 'name': record.field['name']} for record in wiretap.output_records]
             assert sorted(records, key=lambda d: str(d['id'])) == sorted(sample_data[30:40], key=lambda d: str(d['id']))
     finally:
-        if sdc_executor.get_pipeline_status(pipeline).response.json().get('status') == 'RUNNING':
+        if (pipeline is not None
+                and sdc_executor.get_pipeline_status(pipeline).response.json().get('status') == 'RUNNING'):
             sdc_executor.stop_pipeline(pipeline=pipeline, force=True)
         database.deactivate_and_drop_replication_slot(replication_slot)
         if table is not None:
@@ -372,6 +374,7 @@ def test_start_not_from_latest(sdc_builder,
                                                sqlalchemy.String(20)))
     replication_slot = get_random_string(string.ascii_lowercase, 10)
 
+    pipeline = None
     try:
         table.create(database.engine)
         if create_slot:
@@ -461,7 +464,8 @@ def test_start_not_from_latest(sdc_builder,
                        for record in wiretap.output_records]
             assert sorted(records, key=lambda d: str(d['name'])) == sorted(sample_data_3, key=lambda d: str(d['name']))
     finally:
-        if sdc_executor.get_pipeline_status(pipeline).response.json().get('status') == 'RUNNING':
+        if (pipeline is not None
+                and sdc_executor.get_pipeline_status(pipeline).response.json().get('status') == 'RUNNING'):
             sdc_executor.stop_pipeline(pipeline=pipeline, force=True)
         database.deactivate_and_drop_replication_slot(replication_slot)
         if table is not None:
@@ -511,6 +515,7 @@ def test_aurora_postgres_cdc_client_basic(sdc_builder,
     sdc_executor.add_pipeline(pipeline)
 
     table = _create_table_in_database(table_name, database)
+    connection = None
     try:
         sdc_executor.start_pipeline(pipeline)
         connection = database.engine.connect()
@@ -578,7 +583,8 @@ def test_aurora_postgres_cdc_client_basic(sdc_builder,
         database.deactivate_and_drop_replication_slot(replication_slot_name)
         if table is not None:
             table.drop(database.engine)
-        connection.close()
+        if connection is not None:
+            connection.close()
 
 
 @database('postgresqlaurora')
@@ -623,6 +629,7 @@ def test_aurora_postgres_cdc_max_poll_attempts(sdc_builder,
     sdc_executor.add_pipeline(pipeline)
 
     table = _create_table_in_database(table_name, database)
+    connection = None
     try:
         sdc_executor.start_pipeline(pipeline)
         connection = database.engine.connect()
@@ -685,7 +692,8 @@ def test_aurora_postgres_cdc_max_poll_attempts(sdc_builder,
         database.deactivate_and_drop_replication_slot(replication_slot_name)
         if table is not None:
             table.drop(database.engine)
-        connection.close()
+        if connection is not None:
+            connection.close()
 
 
 @database('postgresqlaurora')
@@ -733,6 +741,7 @@ def test_aurora_postgres_cdc_client_filtering_table(sdc_builder,
 
     table_allow = _create_table_in_database(table_name_allow, database)
     table_deny = _create_table_in_database(table_name_deny, database)
+    connection = None
     try:
         sdc_executor.start_pipeline(pipeline)
         connection = database.engine.connect()
@@ -807,7 +816,8 @@ def test_aurora_postgres_cdc_client_filtering_table(sdc_builder,
             table_allow.drop(database.engine)
         if table_deny is not None:
             table_deny.drop(database.engine)
-        connection.close()
+        if connection is not None:
+            connection.close()
 
 
 @database('postgresqlaurora')
@@ -857,6 +867,7 @@ def test_aurora_postgres_cdc_client_remove_replication_slot(sdc_builder,
     sdc_executor.add_pipeline(pipeline)
 
     table = _create_table_in_database(table_name, database)
+    connection = None
     try:
         sdc_executor.start_pipeline(pipeline)
         connection = database.engine.connect()
@@ -876,7 +887,8 @@ def test_aurora_postgres_cdc_client_remove_replication_slot(sdc_builder,
             sdc_executor.stop_pipeline(pipeline=pipeline, force=True)
         if table is not None:
             table.drop(database.engine)
-        connection.close()
+        if connection is not None:
+            connection.close()
 
 
 @database('postgresqlaurora')
@@ -940,6 +952,7 @@ def test_aurora_postgres_cdc_client_multiple_concurrent_operations(sdc_builder,
     sdc_executor.add_pipeline(pipeline)
 
     table = _create_table_in_database(table_name, database)
+    connections = None
     try:
         pipeline_cmd = sdc_executor.start_pipeline(pipeline)
         connections = [database.engine.connect() for _ in range(threads_count)]
@@ -1034,8 +1047,9 @@ def test_aurora_postgres_cdc_client_multiple_concurrent_operations(sdc_builder,
         database.deactivate_and_drop_replication_slot(replication_slot_name)
         if table is not None:
             table.drop(database.engine)
-        for conn in connections:
-            conn.close()
+        if connections is not None:
+            for conn in connections:
+                conn.close()
 
 
 @database('postgresqlaurora')
@@ -1204,6 +1218,7 @@ def test_aurora_postgres_cdc_wal_sender_status_metrics(sdc_builder,
     pipeline = pipeline_builder.build().configure_for_environment(database)
     sdc_executor.add_pipeline(pipeline)
     table = _create_table_in_database(table_name, database)
+    connection = None
     try:
         sdc_executor.start_pipeline(pipeline)
         connection = database.engine.connect()
@@ -1233,7 +1248,8 @@ def test_aurora_postgres_cdc_wal_sender_status_metrics(sdc_builder,
             sdc_executor.stop_pipeline(pipeline=pipeline, force=True)
         database.deactivate_and_drop_replication_slot(replication_slot_name)
         table.drop(database.engine)
-        connection.close()
+        if connection is not None:
+            connection.close()
 
 
 @database('postgresqlaurora')
@@ -1282,6 +1298,7 @@ def test_aurora_postgres_cdc_queue_buffering_metrics(sdc_builder,
     sdc_executor.add_pipeline(pipeline)
 
     tables = [_create_table_in_database(table_name, database) for table_name in table_names]
+    connection = None
     try:
         sdc_executor.start_pipeline(pipeline)
         connection = database.engine.connect()
@@ -1320,7 +1337,8 @@ def test_aurora_postgres_cdc_queue_buffering_metrics(sdc_builder,
         database.deactivate_and_drop_replication_slot(replication_slot_name)
         for t in tables:
             t.drop(database.engine)
-        connection.close()
+        if connection is not None:
+            connection.close()
 
 
 @database('postgresqlaurora')
@@ -1374,6 +1392,7 @@ def test_aurora_postgres_cdc_ssl_enabled(sdc_builder,
     sdc_executor.add_pipeline(pipeline)
 
     table = _create_table_in_database(table_name, database)
+    connection = None
     try:
         sdc_executor.start_pipeline(pipeline)
         connection = database.engine.connect()
@@ -1443,7 +1462,8 @@ def test_aurora_postgres_cdc_ssl_enabled(sdc_builder,
         # database.deactivate_and_drop_replication_slot(replication_slot_name)
         if table is not None:
             table.drop(database.engine)
-        connection.close()
+        if connection is not None:
+            connection.close()
 
 
 @database('postgresqlaurora')
@@ -1460,14 +1480,12 @@ def test_aurora_postgres_cdc_client_primary_keys_metadata_headers(sdc_builder,
                                                                   record_contents,
                                                                   parse_datetimes):
     pipeline = None
-
+    database_connection = None
+    replication_slot_name = get_random_string(string.ascii_lowercase, 10)
+    table_name = f'stf_{get_random_string(string.ascii_lowercase, 16)}'
     try:
 
         database_connection = database.engine.connect()
-
-        replication_slot_name = get_random_string(string.ascii_lowercase, 10)
-
-        table_name = f'stf_{get_random_string(string.ascii_lowercase, 16)}'
 
         pipeline_builder = sdc_builder.get_pipeline_builder()
         aurora_postgres_cdc_client = pipeline_builder.add_stage('Aurora PostgreSQL CDC Client')
@@ -1608,7 +1626,8 @@ def test_aurora_postgres_cdc_client_primary_keys_metadata_headers(sdc_builder,
             sdc_executor.stop_pipeline(pipeline=pipeline, force=True)
         database.deactivate_and_drop_replication_slot(replication_slot_name)
         try:
-            database_connection.execute(f'drop table {table_name}')
+            if database_connection is not None:
+                database_connection.execute(f'drop table {table_name}')
         except:
             pass
         database_connection.close()
@@ -1628,13 +1647,14 @@ def test_aurora_postgres_cdc_client_primary_keys_headers(sdc_builder,
     Test to check all headers for primary keys are present in the output records.
     """
 
+    pipeline = None
+    table = None
+    database_connection = None
+    replication_slot_name = get_random_string(string.ascii_lowercase, 10)
+    table_name = f'stf_{get_random_string(string.ascii_lowercase, 16)}'
     try:
 
         database_connection = database.engine.connect()
-
-        replication_slot_name = get_random_string(string.ascii_lowercase, 10)
-
-        table_name = f'stf_{get_random_string(string.ascii_lowercase, 16)}'
 
         pipeline_builder = sdc_builder.get_pipeline_builder()
         postgres_cdc_client = pipeline_builder.add_stage('Aurora PostgreSQL CDC Client')
@@ -1817,6 +1837,7 @@ def test_aurora_postgres_cdc_client_primary_keys_headers(sdc_builder,
             pass
 
         try:
-            database_connection.close()
+            if database_connection is not None:
+                database_connection.close()
         except:
             pass
